Refactor coordinate detector
