--- django_notate/sync_files/coordinates.py
@@ -1,31 +1,0 @@
-# import the necessary packages
-from __future__ import print_function
-from imutils import perspective
-from imutils import contours
-import numpy as np
-import argparse
-import imutils
-import cv2
- 
-def order_points_old(pts):
-	# initialize a list of coordinates that will be ordered
-	# such that the first entry in the list is the top-left,
-	# the second entry is the top-right, the third is the
-	# bottom-right, and the fourth is the bottom-left
-	rect = np.zeros((4, 2), dtype="float32")
- 
-	# the top-left point will have the smallest sum, whereas
-	# the bottom-right point will have the largest sum
-	s = pts.sum(axis=1)
-	rect[0] = pts[np.argmin(s)]
-	rect[2] = pts[np.argmax(s)]
- 
-	# now, compute the difference between the points, the
-	# top-right point will have the smallest difference,
-	# whereas the bottom-left will have the largest difference
-	diff = np.diff(pts, axis=1)
-	rect[1] = pts[np.argmin(diff)]
-	rect[3] = pts[np.argmax(diff)]
- 
-	# return the ordered coordinates
-	return rect
