+++ django_notate/sync_files/coordinates_new.py
@@ -1,7 +1,5 @@
 # import the necessary packages
 from scipy.spatial import distance as dist
-from coordinates import order_points_old
-# from __future__ import print_function
 from imutils import perspective
 from imutils import contours
 import numpy as np
@@ -9,35 +7,30 @@ import cv2
 import argparse
 import imutils
 
-def order_points(pts):
-	# sort the points based on their x-coordinates
-	xSorted = pts[np.argsort(pts[:, 0]), :]
-
-	# grab the left-most and right-most points from the sorted
-	# x-roodinate points
-	leftMost = xSorted[:2, :]
-	rightMost = xSorted[2:, :]
-
-	# now, sort the left-most coordinates according to their
-	# y-coordinates so we can grab the top-left and bottom-left
-	# points, respectively
-	leftMost = leftMost[np.argsort(leftMost[:, 1]), :]
-	(tl, bl) = leftMost
-
-	# now that we have the top-left coordinate, use it as an
-	# anchor to calculate the Euclidean distance between the
-	# top-left and right-most points; by the Pythagorean
-	# theorem, the point with the largest distance will be
-	# our bottom-right point
-	D = dist.cdist(tl[np.newaxis], rightMost, "euclidean")[0]
-	(br, tr) = rightMost[np.argsort(D)[::-1], :]
-
-	# return the coordinates in top-left, top-right,
-	# bottom-right, and bottom-left order
-	return np.array([tl, tr, br, bl], dtype="float32")
+def order_points_old(pts):
+	# initialize a list of coordinates that will be ordered
+	# such that the first entry in the list is the top-left,
+	# the second entry is the top-right, the third is the
+	# bottom-right, and the fourth is the bottom-left
+	rect = np.zeros((4, 2), dtype="float32")
+ 
+	# the top-left point will have the smallest sum, whereas
+	# the bottom-right point will have the largest sum
+	s = pts.sum(axis=1)
+	rect[0] = pts[np.argmin(s)]
+	rect[2] = pts[np.argmax(s)]
+ 
+	# now, compute the difference between the points, the
+	# top-right point will have the smallest difference,
+	# whereas the bottom-left will have the largest difference
+	diff = np.diff(pts, axis=1)
+	rect[1] = pts[np.argmin(diff)]
+	rect[3] = pts[np.argmax(diff)]
+ 
+	# return the ordered coordinates
+	return rect
 
 def coordinates(new_order):
-    # print "made it to coordinates"
 
 	# construct the argument parse and parse the arguments
 	# TODO: Remove argument parsing from here
@@ -45,8 +38,6 @@ def coordinates(new_order):
 	#ap.add_argument("-n", "--new", type=int, default=-1,
     #		help="whether or not the new order points should should be used")
 	#args = vars(ap.parse_args());
-
-    # print "made it to coordinates"
 
 	# load our input image, convert it to grayscale, and blur it slightly
 	image = cv2.imread("DarkImageMove.jpg")
@@ -85,19 +76,12 @@ def coordinates(new_order):
 
 		# show the original coordinates
 		# print("Object #{}:".format(i + 1))
-		# print "woah"
-		# print(box)
 
 	# order the points in the contour such that they appear
 		# in top-left, top-right, bottom-right, and bottom-left
 		# order, then draw the outline of the rotated bounding
 		# box
 		rect = order_points_old(box)
-
-		# check to see if the new method should be used for
-		# ordering the coordinates
-		if new_order > 0:
-			rect = perspective.order_points(box)
 
 		# show the re-ordered coordinates
 		# print(rect.astype("int"))
@@ -117,7 +101,5 @@ def coordinates(new_order):
 		# cv2.waitKey(0)
 
 		# return (rect.astype("int"))
-		# print "box"
-		# print box
 		return box
